select account trigger - move to separate file
