+++ packages/renderer/src/screens/Home/Send/Dialogs/DialogSelectAccount.tsx
@@ -10,13 +10,13 @@ import { useEffect, useState } from 'react';
 import { checkAddress } from '@polkadot/util-crypto';
 import { Identicon } from '@polkadot-live/ui/components';
 import { ellipsisFn } from '@w3ux/utils';
+import { TriggerSelectAccount } from '.';
 import {
   AddressesWrapper,
   AddressItem,
   ConfirmBtn,
   InputIdenticonWrapper,
   InputWrapper,
-  SelectedAddressItem,
   TriggerButton,
 } from './Wrappers';
 
@@ -25,68 +25,6 @@ import type { ChangeEvent } from 'react';
 import type { ChainID } from '@polkadot-live/types/chains';
 import type { SendRecipient } from '../types';
 import type { SendAccount } from '@polkadot-live/types/accounts';
-
-interface DynamicTriggerProps {
-  accountRole: 'recipient' | 'sender';
-  recipient: SendRecipient | null;
-  sender: SendAccount | null;
-}
-
-const DynamicTrigger = ({
-  accountRole,
-  recipient,
-  sender,
-}: DynamicTriggerProps) => {
-  const { getTheme, getOnlineMode } = useConnections();
-  const theme = getTheme();
-
-  let sendAccount: SendAccount | SendRecipient | null;
-  let accountNotSet: boolean;
-  let notSetMessage: string;
-  let label = '';
-
-  switch (accountRole) {
-    case 'recipient': {
-      sendAccount = recipient;
-      accountNotSet = recipient === null || recipient?.address === '';
-      notSetMessage = 'Select Recipient';
-
-      if (recipient) {
-        const { accountName, address } = recipient;
-        label = accountName !== null ? accountName : ellipsisFn(address, 5);
-      }
-      break;
-    }
-    case 'sender': {
-      sendAccount = sender;
-      accountNotSet = sender === null;
-      notSetMessage = 'Select Sender';
-
-      if (sender) {
-        label = sender.alias;
-      }
-      break;
-    }
-  }
-
-  return accountNotSet ? (
-    <span style={{ textAlign: 'left', flex: 1 }}>{notSetMessage}</span>
-  ) : (
-    <SelectedAddressItem
-      className={!getOnlineMode() ? 'disable' : ''}
-      $theme={theme}
-    >
-      <Styles.FlexRow $gap={'1.25rem'} style={{ width: '100%' }}>
-        <div className="identicon" style={{ minWidth: 'fit-content' }}>
-          <Identicon value={sendAccount!.address} fontSize="1.9rem" />
-        </div>
-        <Styles.FlexColumn $rowGap={'0.5rem'} style={{ flex: 1, minWidth: 0 }}>
-          <h3>{label}</h3>
-        </Styles.FlexColumn>
-      </Styles.FlexRow>
-    </SelectedAddressItem>
-  );
-};
 
 export const DialogSelectAccount = ({
   accounts,
@@ -279,8 +217,7 @@ export const DialogSelectAccount = ({
           $theme={theme}
           className={!getOnlineMode() ? 'disable' : ''}
         >
-          {/** Render Recipient or Sender Trigger */}
-          <DynamicTrigger
+          <TriggerSelectAccount
             accountRole={accountRole}
             recipient={recipient}
             sender={sender}
